+++ Examples/UIKit-ObjC/Superwall-UIKit-ObjC/Services/SSASuperwallService.m
@@ -106,20 +106,12 @@ static inline SWKPurchaseResult SWKPurchaseResultFromTransactionState(SKPaymentT
   [[Superwall sharedInstance] setDelegate:self];
 }
 
-- (void)logInWithCompletion:(nullable void (^)(void))completion {
-  [[Superwall sharedInstance] identifyWithUserId:kDemoAPIKey completion:^(NSError * _Nullable error) {
-    if (error != nil) {
-      NSLog(@"%@", error.localizedDescription);
-    }
-
-    if (completion) {
-      completion();
-    }
-  }];
+- (void)logIn {
+  [[Superwall sharedInstance] identifyWithUserId:kDemoUserId];
 }
 
-- (void)logOutWithCompletion:(nullable void (^)(void))completion {
-  [[Superwall sharedInstance] resetWithCompletion:completion];
+- (void)logOut {
+  [[Superwall sharedInstance] reset];
 }
 
 - (void)handleDeepLinkWithURL:(NSURL *)URL {