+++ Sources/SuperwallKit/Config/ConfigManager.swift
@@ -25,12 +25,38 @@ class ConfigManager {
   var options = SuperwallOptions()
 
   /// A dictionary of triggers by their event name.
-  var triggersByEventName: [String: Trigger] = [:]
+  var triggersByEventName: [String: Trigger] {
+    get {
+      queue.sync { [unowned self] in
+        self._triggersByEventName
+      }
+    }
+    set {
+      queue.async { [unowned self] in
+        self._triggersByEventName = newValue
+      }
+    }
+  }
+  private var _triggersByEventName: [String: Trigger] = [:]
 
   /// A memory store of assignments that are yet to be confirmed.
   ///
   /// When the trigger is fired, the assignment is confirmed and stored to disk.
-  var unconfirmedAssignments: [Experiment.ID: Experiment.Variant] = [:]
+  var unconfirmedAssignments: [Experiment.ID: Experiment.Variant] {
+    get {
+      queue.sync { [unowned self] in
+        self._unconfirmedAssignments
+      }
+    }
+    set {
+      queue.async { [unowned self] in
+        self._unconfirmedAssignments = newValue
+      }
+    }
+  }
+  private var _unconfirmedAssignments: [Experiment.ID: Experiment.Variant] = [:]
+
+  private let queue = DispatchQueue(label: "com.superwall.configmanager")
 
   private unowned let storeKitManager: StoreKitManager
   private unowned let storage: Storage
@@ -239,7 +265,11 @@ class ConfigManager {
           responseIdentifiers: .init(paywallId: identifier),
           overrides: nil
         )
-        _ = try? await paywallManager.getPaywallViewController(from: request)
+        _ = try? await paywallManager.getPaywallViewController(
+          from: request,
+          isPreloading: true,
+          isDebuggerLaunched: false
+        )
       }
     }
   }