+++ Sources/SuperwallKit/Paywall/Manager/PaywallManager.swift
@@ -9,36 +9,32 @@ import Foundation
 import UIKit
 
 class PaywallManager {
-  @MainActor
   var presentedViewController: PaywallViewController? {
-    return PaywallViewController.cache.first { $0.isActive }
+    return cache.getActivePaywallViewController()
 	}
   private unowned let paywallRequestManager: PaywallRequestManager
-  private unowned let factory: ViewControllerFactory & CacheFactory
+  private unowned let factory: ViewControllerFactory & CacheFactory & DeviceInfoFactory
 
-  private lazy var cache: PaywallCache = factory.makeCache()
+  private lazy var cache: PaywallViewControllerCache = factory.makeCache()
 
   init(
-    factory: ViewControllerFactory & CacheFactory,
+    factory: ViewControllerFactory & CacheFactory & DeviceInfoFactory,
     paywallRequestManager: PaywallRequestManager
   ) {
     self.factory = factory
     self.paywallRequestManager = paywallRequestManager
   }
 
-  @MainActor
-	func removePaywallViewController(identifier: String?) {
-    cache.removePaywallViewController(identifier: identifier)
+  func setActivePaywallVcKey(to key: String) {
+    cache.activePaywallVcKey = key
+  }
+
+	func removePaywallViewController(forKey: String) {
+    cache.removePaywallViewController(forKey: forKey)
 	}
 
-  @MainActor
-	func removePaywallViewController(_ viewController: PaywallViewController) {
-    cache.removePaywallViewController(viewController)
-	}
-
-  @MainActor
 	func resetCache() {
-		cache.clearCache()
+		cache.removeAll()
 	}
 
   /// First, this gets the paywall response for a specified paywall identifier or trigger event.
@@ -53,24 +49,29 @@ class PaywallManager {
   @MainActor
   func getPaywallViewController(
     from request: PaywallRequest,
-    cached: Bool = true
+    isPreloading: Bool,
+    isDebuggerLaunched: Bool
   ) async throws -> PaywallViewController {
     let paywall = try await paywallRequestManager.getPaywall(from: request)
+    let deviceInfo = factory.makeDeviceInfo()
+    let cacheKey = PaywallCacheLogic.key(
+      identifier: paywall.identifier,
+      locale: deviceInfo.locale
+    )
 
-    if cached,
-      let viewController = self.cache.getPaywallViewController(identifier: paywall.identifier) {
-      // Set product-related vars again incase products have been substituted into paywall.
-      viewController.paywall.products = paywall.products
-      viewController.paywall.productIds = paywall.productIds
-      viewController.paywall.swProducts = paywall.swProducts
-      viewController.paywall.productVariables = paywall.productVariables
-      viewController.paywall.swProductVariablesTemplate = paywall.swProductVariablesTemplate
-      viewController.paywall.isFreeTrialAvailable = paywall.isFreeTrialAvailable
-      viewController.paywall.productsLoadingInfo = paywall.productsLoadingInfo
+    if !isDebuggerLaunched,
+      let viewController = self.cache.getPaywallViewController(forKey: cacheKey) {
+      if !isPreloading {
+        viewController.paywall.overrideProductsIfNeeded(from: paywall)
+      }
       return viewController
     }
 
-    let paywallViewController = factory.makePaywallViewController(for: paywall)
+    let paywallViewController = factory.makePaywallViewController(
+      for: paywall,
+      withCache: cache
+    )
+    cache.save(paywallViewController, forKey: cacheKey)
 
     if let window = UIApplication.shared.activeWindow {
       paywallViewController.view.alpha = 0.01
@@ -83,5 +84,5 @@ class PaywallManager {
     }
 
     return paywallViewController
-	}
+  }
 }