+++ Sources/SuperwallKit/Paywall/Request/Operators/AddProductsOperator.swift
@@ -53,7 +53,6 @@ extension AnyPublisher where Output == PipelineData, Failure == Error {
         paywall.productVariables = outcome.productVariables
         paywall.swProductVariablesTemplate = outcome.swProductVariablesTemplate
         paywall.isFreeTrialAvailable = outcome.isFreeTrialAvailable
-
         paywall.productsLoadingInfo.endAt = Date()
 
         return (paywall, input.request)