+++ CHANGELOG.md
@@ -6,15 +6,22 @@ The changelog for `SuperwallKit`. Also see the [releases](https://github.com/sup
 
 ### Breaking Changes
 
-- The non-async version of `identify` was never throwing an error due to the async error of the underlying code. This now passes the error via its completion block.
+- `identify(userId:)` is not longer a throwing async function. Any error that occurs is logged.
+- `reset` is no longer an async function.
+- `presentedViewController` and `latestPaywallInfo` no longer restricted to the main actor.
 - Removes `localizationOverride(localeIdentifier:)` and replaces it with the `SuperwallOption` `localeIdentifier`. You set this on configure.
 - `setUserAttributes(_:)` is now an async function. It always was async under the hood but this makes it more explicit and helps you avoid race conditions.
 - Removes delegate from `configure`. You now set the delegate via `Superwall.shared.delegate`.
 - Removes `presenter` introduced in beta 5.
 
+### Enhancements
+
+- You can now target `device.isSandbox` in rules.
+
 ### Fixes
 
 - Fixes bug where calling identify and immediately tracking a paywall would result in an error if it happened before configure returned.
+- Fixes compiler bug when calling track.
 
 ## 3.0.0-beta.5